Code rearragement

+++ kernel/include/vexos/lib/math.h
@@ -15,7 +15,7 @@
 #define M_SQRT2	    1.41421356237309504880	/* sqrt(2) */
 #define M_SQRT1_2	0.70710678118654752440	/* 1/sqrt(2) */
 
-float cos(float x);
-float sin(float x);
+double cos(double x);
+double sin(double x);
 
 #endif
